Сделает рефакторинг: выделит связанные компоненты из MainContent
